+++ src/ambition/version_details.cc
@@ -21,6 +21,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("When selecting where to place a building, show grid rules of the building, and grid rules one tile out, for a walkable perimeter, to allow for easier placement and planning of buildings."),
           _("When selecting where to place a building, highlight valid locations, darken invalid locations, and shade un-walkable locations red, to create a guide for placing and planning buildings.  This is especially helpful with Harbours."),
+          _("Highlight possible Harbour locations brighter when they can link to other buildings, to allow for easier placement of buildings."),
           _("Show the per-unit and total costs, along with the percentages of the fixed expense, for units on the Military Report."),
           _("Show a total for the unit costs, along with the percentage of fixed expense, on the Military Report."),
           _("Added a keybind for confirming in saving and loading menus (defaults to Enter), settable in config.txt with `bindkey = KEYEVENT_ _CONFIRM,Return`."),